Reorganize game/ file structure
