+++ game/reducers/board.jsx
