+++ game/reducers/index.jsx
