+++ game/reducers/player.jsx
